feat: 입력한 금액에 따라 로또 개수 출력
로또 개수 출력 및 상수 선언
