+++ src/view/output.js
@@ -1,5 +1,10 @@
 import { Console } from "@woowacourse/mission-utils";
+import { OUTPUT_MESSAGES } from "../constants/Message.js";
 
 export const printError = (errorMessage) => {
-	return Console.readLineAsync(errorMessage);
+	Console.print(errorMessage);
+};
+
+export const printLottoCount = (number) => {
+	Console.print(OUTPUT_MESSAGES.OUTPUT_LOTTO_COUNT(number));
 };
